heavily refactor api module
Update the API module to use the new storage interface. Due to the way
we defined the storage interfaces I was able to greatly reduce the
amount of code here. I made the resource interfaces more modular and
so moved most things into the __init__ since they were not
individually worthy of their own file