Fix incorrect scheme bug
Depending on the database we would get different results. Sqlite3 returns the
correct scheme when no kwargs are specified however on Postgres it returns the
last created one. Setting an order_by caused more issues as Postgres would then
return a non-default when kwargs were specified but a default was required, this
conditionally adds the union if the consumer is looking for a specific but a
default if not. Tested on Sqlite3 and Postgres