Notifications are now supported
Implemented notifications using the django-notifications app found at:
https://github.com/django-notifications/django-notifications

This gives the user a stream of activity that is relevant to them by being a
watcher or involved with a ticket. This stream is available on their dashboard.
Additionally projects now show an activity stream on their show page.

There were various minor bug fixes included in this work as well as updating
tests.

Next Steps:

- Update the project page to allow easy access to "open tickets" since that info
  is no longer displayed on the front page for a project.

- Allow users to become watchers of a ticket and automatically make them a
  watcher on commenting.