Project Administration
This commit adds the appropriate screens and routes for performing
project administration according to this issue:
https://github.com/praelatus/praelatus/issues/152.

This allows Project Leads and other Administrators to properly control
access to their projects, control what ticket types are available, and
control how workflows are attached to those ticket types. This does not
however allow for FieldSchemes to be managed by project admins. We still
have some documentation and discovery to do around that before I would
feel safe writing any kind of UI for it but this is ready to go so we
should go ahead and cut it into the release.

Additionally this does a major UX/UI overhaul changing the primary color
of the site to match the new purple we are using on the logo and also
cleans up the ticket show screen etc to look cleaner, clearer, and under
control.

Bugs that were fixed in this commit:
 - Queries ignore user permissions (https://github.com/praelatus/praelatus/issues/156)
 - Permissions didn't fallback to anonymous when the user didn't have
 permission, now they do.